corregir detalles causados por merge de dev a landing-page
Se corrigieron algunos detalles al haber renombrado el archivo principal. Se cambiaron las url's para direccionar correctamente a la carpeta global de "assets".
