+++ dream-lab-frontend/src/pages/LandingPage/components/Footer/Footer.jsx
@@ -3,11 +3,11 @@ import "./Footer.css";
 import tecLogo from "src/assets/LandingPage/tecnologico-de-monterrey-white.png";
 
 function Footer() {
-    return (
-        <div className="footer-container">
-            <img className="tec-logo" src={tecLogo} />
-        </div>
-    );
+  return (
+    <div className="footer-container">
+      <img className="tec-logo" src={tecLogo} />
+    </div>
+  );
 }
 
 export default Footer;
