+++ dream-lab-frontend/src/pages/LandingPage/components/InfoCard/InfoCard.css
@@ -1,12 +1,12 @@
 @import url('https://fonts.googleapis.com/css2?family=Poppins:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap');
 
 .info-card-container {
-    display: flex;
-    flex-flow: row wrap;
     /* background-color: rgba(255, 255, 255, 0.05); */
-    padding: 3rem;
     /* border-radius: 4rem; */
     /* border: 1px solid white; */
+    display: flex;
+    flex-flow: row wrap;
+    padding: 3rem;
 }
 
 .title {
