+++ dream-lab-frontend/src/pages/LandingPage/components/Title/Title.css
@@ -72,24 +72,28 @@
     text-transform: uppercase;
 }
 
+.landing-down-arrow {
+    cursor: pointer;
+}
+
 @media (max-width: 820px) {
 .button-more {
-        width: 100%;
-        height: 5vh;
-        border: 0px solid white;
-        border-top: 2px solid white;
-        border-radius: 0px;
-        border-bottom-right-radius: 6vw;
-        border-bottom-left-radius: 6vw;
-        font-size: small;
+    width: 100%;
+    height: 5vh;
+    border: 0px solid white;
+    border-top: 2px solid white;
+    border-radius: 0px;
+    border-bottom-right-radius: 6vw;
+    border-bottom-left-radius: 6vw;
+    font-size: small;
     }
 
 .button-more:hover {
-        width: 100%;
-        border: 0px solid white;
-        border-bottom: 2px solid white;
-        border-radius: 0px;
-        border-bottom-right-radius: 5vw;
-        border-bottom-left-radius: 5vw;
+    width: 100%;
+    border: 0px solid white;
+    border-bottom: 2px solid white;
+    border-radius: 0px;
+    border-bottom-right-radius: 5vw;
+    border-bottom-left-radius: 5vw;
     }
 }
